fromACL method and acl module require
