+++ mod/user/acl.js
@@ -2,37 +2,30 @@ const { Pool } = require('pg');
 
 const connection = process.env.PRIVATE?.split('|') || process.env.PUBLIC?.split('|')
 
-let pool = null
+if(!connection || !connection[1]) return
 
-module.exports = () => {
+const acl_table = connection[1].split('.').pop()
 
-  if(!connection || !connection[1]) return
+const acl_schema = connection[1].split('.')[0] === acl_table ? 'public' : connection[1].split('.')[0]
 
-  const acl_table = connection[1].split('.').pop()
+const pool = new Pool({
+  connectionString: connection[0]
+})
 
-  const acl_schema = connection[1].split('.')[0] === acl_table ? 'public' : connection[1].split('.')[0]
+module.exports = async (q, arr) => {
 
-  // Create PostgreSQL connection pool for ACL table.
-  pool ??= new Pool({
-    connectionString: connection[0]
-  })
+  try {
 
-  // Method to query ACL. arr must be empty array by default.
-  return async (q, arr) => {
+    const client = await pool.connect()
 
-    try {
+    const { rows } = await client.query(q.replace(/acl_table/g, acl_table).replace(/acl_schema/g, acl_schema), arr)
 
-      const client = await pool.connect()
+    client.release()
 
-      const { rows } = await client.query(q.replace(/acl_table/g, acl_table).replace(/acl_schema/g, acl_schema), arr)
+    return rows
 
-      client.release()
-      
-      return rows
-    
-    } catch (err) {
-      console.error(err)
-      return err
-    }
+  } catch (err) {
+    console.error(err)
+    return err
   }
 }
