+++ mod/user/register.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcryptjs')
 
 const crypto = require('crypto')
 
-const acl = require('./acl')()
+const acl = require('./acl')
 
 const mailer = require('../utils/mailer')
